network/openai: split request and response adapters from transport
