+++ mesh-llm/src/network/openai/mod.rs
@@ -2,5 +2,7 @@ pub(crate) mod adapter;
 pub(crate) mod backend;
 pub(crate) mod errors;
 pub(crate) mod ingress;
+pub(crate) mod request_adapter;
+pub(crate) mod response_adapter;
 pub(crate) mod schema;
 pub(crate) mod transport;
